# Change Set to implement Hash directly to reduce memory usage.