--- a/src/set.cr
+++ b/src/set.cr
@@ -2,8 +2,12 @@ class Set(T)
   include Enumerable(T)
   include Iterable
 
+  getter length
+
   def initialize
-    @hash = Hash(T, Nil).new
+    @buckets = Pointer(Entry(T)?).malloc(11)
+    @buckets_length = 11
+    @length = 0
   end
 
   def self.new(array : Array(T))
@@ -19,7 +23,22 @@ class Set(T)
   end
 
   def add(object : T)
-    @hash[object] = nil
+    rehash if @length > 5 * @buckets_length
+
+    index = bucket_index object
+    entry = insert_in_bucket index, object
+    return nil unless entry
+
+    @length += 1
+
+    if last = @last
+      last.fore = entry
+      entry.back = last
+    end
+
+    @last = entry
+    @first = entry unless @first
+    nil
   end
 
   def merge(elems)
@@ -27,68 +46,189 @@ class Set(T)
   end
 
   def includes?(object)
-    @hash.has_key?(object)
+    !!find_entry(object)
+  end
+
+  protected def find_entry(object)
+    index = bucket_index object
+    entry = @buckets[index]
+    find_entry_in_bucket entry, object
+  end
+
+  private def find_entry_in_bucket(entry, object)
+    while entry
+      if object == entry.object
+        return entry
+      end
+      entry = entry.next
+    end
+    nil
+  end
+
+  private def insert_in_bucket(index, object)
+    entry = @buckets[index]
+    if entry
+      while entry
+        if object == entry.object
+          return nil
+        end
+        if entry.next
+          entry = entry.next
+        else
+          return entry.next = Entry(T).new(object)
+        end
+      end
+    else
+      return @buckets[index] = Entry(T).new(object)
+    end
+  end
+
+  private def insert_in_bucket_end(index, existing_entry)
+    entry = @buckets[index]
+    if entry
+      while entry
+        if entry.next
+          entry = entry.next
+        else
+          return entry.next = existing_entry
+        end
+      end
+    else
+      @buckets[index] = existing_entry
+    end
+  end
+
+  private def bucket_index(object)
+    (object.hash.abs % @buckets_length).to_i
+  end
+
+  private def calculate_new_size(size)
+    new_size = 8
+    HASH_PRIMES.each do |hash_size|
+      return hash_size if new_size > size
+      new_size <<= 1
+    end
+    raise "Hash table too big"
   end
 
   def delete(object)
-    @hash.delete(object)
+    index = bucket_index(object)
+    entry = @buckets[index]
+
+    previous_entry = nil
+    while entry
+      if object == entry.object
+        back_entry = entry.back
+        fore_entry = entry.fore
+        if fore_entry
+          if back_entry
+            back_entry.fore = fore_entry
+            fore_entry.back = back_entry
+          else
+            @first = fore_entry
+            fore_entry.back = nil
+          end
+        else
+          if back_entry
+            back_entry.fore = nil
+            @last = back_entry
+          else
+            @first = nil
+            @last = nil
+          end
+        end
+        if previous_entry
+          previous_entry.next = entry.next
+        else
+          @buckets[index] = entry.next
+        end
+        @length -= 1
+        return nil
+      end
+      previous_entry = entry
+      entry = entry.next
+    end
+    nil
   end
 
-  def length
-    @hash.length
+  def rehash
+    new_size = calculate_new_size(@length)
+    @buckets = @buckets.realloc(new_size)
+    new_size.times { |i| @buckets[i] = nil }
+    @buckets_length = new_size
+    entry = @first
+    while entry
+      entry.next = nil
+      index = bucket_index entry.object
+      insert_in_bucket_end index, entry
+      entry = entry.fore
+    end
   end
 
   def size
-    length
+    @length
   end
 
   def clear
-    @hash.clear
+    @buckets_length.times do |i|
+      @buckets[i] = nil
+    end
+    @length = 0
+    @first = nil
+    @last = nil
     self
   end
 
   def empty?
-    @hash.empty?
+    @length == 0
   end
 
   def each
-    @hash.each_key do |key|
-      yield key
+    current = @first
+    while current
+      yield current.object
+      current = current.fore
     end
     self
   end
 
   def each
-    @hash.each_key
+    ObjectIterator(T).new(self, @first)
   end
 
   def &(other : Set)
     set = Set(T).new
-    each do |value|
-      set.add value if other.includes?(value)
+    each do |object|
+      set.add object if other.includes?(object)
     end
     set
   end
 
   def |(other : Set(U))
     set = Set(T | U).new
-    each { |value| set.add value }
-    other.each { |value| set.add value }
+    each { |object| set.add object }
+    other.each { |object| set.add object }
     set
   end
 
   def ==(other : Set)
-    same?(other) || @hash == other.@hash
+    return false unless length == other.length
+    each do |object|
+      return false unless other.find_entry(object)
+    end
+    true
   end
 
   def dup
     set = Set(T).new
-    each { |value| set.add value }
+    each { |object| set.add object }
     set
   end
 
   def to_a
-    @hash.keys
+    objects = Array(T).new(@length)
+    each { |object| objects << object }
+    objects
   end
 
   def inspect(io)
@@ -96,7 +236,11 @@ class Set(T)
   end
 
   def hash
-    @hash.hash
+    hash = length
+    each do |object|
+      hash = 31 * hash + object.hash
+    end
+    hash
   end
 
   # Returns true if the set and the given set have at least one
@@ -122,13 +266,86 @@ class Set(T)
 
   def subset?(other : Set)
     return false if other.length < length
-    all? { |value| other.includes?(value) }
+    all? { |object| other.includes?(object) }
   end
 
   def superset?(other : Set)
     return false if other.length > length
-    other.all? { |value| includes?(value) }
+    other.all? { |object| includes?(object) }
   end
+
+  # :nodoc:
+  class ObjectIterator(T)
+    include Iterator(T)
+
+    def initialize(@set, @current)
+    end
+
+    def next
+      if current = @current
+        object = current.object
+        @current = current.fore
+        object
+      else
+        stop
+      end
+    end
+
+    def rewind
+      @current = @set.@first
+    end
+  end
+
+  # :nodoc:
+  class Entry(T)
+    getter :object
+
+    # Next in the linked list of each bucket
+    property :next
+
+    # Next in the ordered sense of hash
+    property :fore
+
+    # Previous in the ordered sense of hash
+    property :back
+
+    def initialize(@object : T)
+    end
+  end
+
+
+  # :nodoc:
+  HASH_PRIMES = [
+    8 + 3,
+    16 + 3,
+    32 + 5,
+    64 + 3,
+    128 + 3,
+    256 + 27,
+    512 + 9,
+    1024 + 9,
+    2048 + 5,
+    4096 + 3,
+    8192 + 27,
+    16384 + 43,
+    32768 + 3,
+    65536 + 45,
+    131072 + 29,
+    262144 + 3,
+    524288 + 21,
+    1048576 + 7,
+    2097152 + 17,
+    4194304 + 15,
+    8388608 + 9,
+    16777216 + 43,
+    33554432 + 35,
+    67108864 + 15,
+    134217728 + 29,
+    268435456 + 3,
+    536870912 + 11,
+    1073741824 + 85,
+    0
+  ]
 end
 
 class Array