I thought it might be a good idea to have a ``Set`` implementation that uses less memory (by not storing a ``nil`` value in each ``Entry``).

A simple performance test shows an advantage (although primarily the advantage should be in less memory used):

```crystal
s = Set(Int32).new
1000.times do
  (0...100000).each do |v|
    s.add v
  end
end
puts s.length
```

Crystal 0.7.4 [42efdbf] (Wed Jun 24 00:55:53 UTC 2015)

```
$ crystal build --release perf_set.cr
$ time ./perf_set 
100000

real	0m0.820s
user	0m0.813s
sys	0m0.008s
```

Crystal master (e19dd68d3151128834e27992f2b84774e242738a)

```
$ bin/crystal build --release perf_set.cr
Using compiled compiler at .build/crystal
$ time ./perf_set 
100000

real	0m0.770s
user	0m0.759s
sys	0m0.008s
```

The parts of ``hash.cr`` that were used were copied directly and modified as necessary to remove the ``value``.