@mrjbq7 2015-07-09
@asterite Wow, awesome quick fix!  It's always frustrating to duplicate code - I didn't know the optimizations worked the way you describe - so very cool.  Thanks for the explanation!